test_resample.py renamed test_base.py
