+++ pandas/tests/resample/test_base.py
